Moved Person into its own file
